Refactor the option mixins and bugfixes (#1)
* Separate popup and toast options

* Refactor el and define popup onClick

* Refactor init() to accept popup and toast options

The original function accpeted only a single options param, which would
contain all options (for both popup and toast mixed). This commit
requires that the two sets of options be provided as separate properties
on the incoming parameter.

* Refactor the default timeout and refresh to 0

* Refactor the refresh logic to work correctly

The relational check was checking for a period shorter than the desired
refresh period, not longer. Also, the code is updated to now set/update
the last shown date every time the modal is shown, and not just on first
display.

* Refactor popup and toast as top-level props

Props added to popAndToast for popup and toast. Options wll now be just
that, options only. The options property is no longer referenced outside
of the init method.

* Correct option overrides for popup and toast

* Remove unused popup prop copy